# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ana Portfolio API
  description: >-
    API do portfólio: conteúdo público, editor autenticado, aparência e
    assistente.


    ### Como usar (passo a passo)

    1. Teste a API em **GET /api/v1/health**.

    2. Leia dados com os **GET** 🔓 (públicos) e anote o campo `revision`.

    3. Para alterar dados, faça login em **POST /api/v1/auth/login** e copie o
    `accessToken`.

    4. Clique em **Authorize** (topo da página), cole o token e confirme.

    5. Use os endpoints 🔒 (POST, PUT, DELETE) enviando `expectedRevision`: a
    **versão atual dos dados**, igual ao `revision` do último GET.


    ### Erros comuns

    - **400**: campo faltando ou em formato errado.

    - **401**: sem login ou token vencido.

    - **404**: o `id` não existe.

    - **409**: `expectedRevision` desatualizado. Faça um novo GET e tente de
    novo.
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    def-2:
      type: object
      description: Texto principal da Home.
      required:
        - title
        - intro
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
          description: Título da Home.
          example: Qualidade de software com método e curiosidade
        intro:
          type: string
          minLength: 1
          maxLength: 3000
          description: Texto de apresentação.
          example: Sou QA e ajudo times a entregar produtos confiáveis.
      title: HeroContent
      example:
        title: Qualidade de software com método e curiosidade
        intro: Sou QA e ajudo times a entregar produtos confiáveis.
    def-3:
      type: object
      description: Texto da seção Sobre.
      required:
        - title
        - body
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
          description: Título da seção.
          example: Sobre mim
        body:
          type: string
          minLength: 1
          maxLength: 5000
          description: Texto completo da seção.
          example: Trabalho com testes funcionais, automação e análise de falhas.
      title: AboutContent
      example:
        title: Sobre mim
        body: Trabalho com testes funcionais, automação e análise de falhas.
...
    def-18:
      type: object
      description: Portfólio completo + controle de versão.
      required:
        - revision
        - updatedAt
        - data
      properties:
        revision:
          type: integer
          minimum: 1
          description: Versão atual dos dados. Use em `expectedRevision` ao salvar.
          example: 5
        updatedAt:
          type: string
          format: date-time
          description: Data da última alteração.
          example: 2025-01-15T12:00:00.000Z
        data:
          $ref: "#/components/schemas/def-17"
      title: PortfolioSnapshot
...
    def-20:
      type: object
      description: "Corpo de escrita: versão atual + os dados novos."
      required:
        - expectedRevision
        - data
      additionalProperties: false
      properties:
        expectedRevision:
          type: integer
          minimum: 1
          description: Versão atual dos dados. Use o valor `revision` retornado pelo
            último GET.
          example: 4
        data:
          $ref: "#/components/schemas/def-2"
      title: HeroWrite
      example:
        expectedRevision: 4
        data:
          title: Qualidade de software com método e curiosidade
          intro: Sou QA e ajudo times a entregar produtos confiáveis.
...
paths:
...
  /api/v1/hero:
    get:
      summary: Mostra o texto da Home
      tags:
        - Content
      description: |-
        🔓 **Público** — não precisa de login.

        Retorna o título e a apresentação da página inicial.
      responses:
        "200":
          description: Conteúdo da Home.
          content:
            application/json:
              schema:
                description: Conteúdo da Home.
                type: object
                required:
                  - revision
                  - updatedAt
                  - data
                properties:
                  revision:
                    type: integer
                    description: Versão atual dos dados. Use em `expectedRevision` ao salvar.
                    example: 4
                  updatedAt:
                    type: string
                    format: date-time
                    description: Data da última alteração.
                    example: 2025-01-15T12:00:00.000Z
                  data:
                    $ref: "#/components/schemas/def-2"
    put:
      summary: Atualiza o texto da Home
      tags:
        - Content
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Troca o título e a apresentação da página inicial. Antes de enviar,
        consulte o GET correspondente e use o valor `revision` retornado em
        `expectedRevision`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/def-20"
      security: *a1
      responses:
        "200":
          description: Salvo. Retorna o portfólio completo com a nova `revision`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-18"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "409":
          description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
            existe). Faça um novo GET e tente de novo."
          content:
            application/json:
              schema:
                description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
                  existe). Faça um novo GET e tente de novo."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: REVISION_CONFLICT
                  message:
                    type: string
                    description: Explicação do erro.
                    example: O portfólio foi alterado depois que esta edição começou. Recarregue a
                      revisão atual antes de salvar.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      expectedRevision: 3
                      currentRevision: 4
  /api/v1/about:
    get:
      summary: Mostra a seção Sobre
      tags:
        - Content
      description: |-
        🔓 **Público** — não precisa de login.

        Retorna o título e o texto da seção Sobre.
      responses:
        "200":
          description: Conteúdo da seção Sobre.
          content:
            application/json:
              schema:
                description: Conteúdo da seção Sobre.
                type: object
                required:
                  - revision
                  - updatedAt
                  - data
                properties:
                  revision:
                    type: integer
                    description: Versão atual dos dados. Use em `expectedRevision` ao salvar.
                    example: 4
                  updatedAt:
                    type: string
                    format: date-time
                    description: Data da última alteração.
                    example: 2025-01-15T12:00:00.000Z
                  data:
                    $ref: "#/components/schemas/def-3"
    put:
      summary: Atualiza a seção Sobre
      tags:
        - Content
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Troca o título e o texto da seção Sobre. Antes de enviar, consulte o GET
        correspondente e use o valor `revision` retornado em `expectedRevision`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/def-21"
      security: *a1
      responses:
        "200":
          description: Salvo. Retorna o portfólio completo com a nova `revision`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-18"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "409":
          description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
            existe). Faça um novo GET e tente de novo."
          content:
            application/json:
              schema:
                description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
                  existe). Faça um novo GET e tente de novo."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: REVISION_CONFLICT
                  message:
                    type: string
                    description: Explicação do erro.
                    example: O portfólio foi alterado depois que esta edição começou. Recarregue a
                      revisão atual antes de salvar.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      expectedRevision: 3
                      currentRevision: 4
  /api/v1/competencies:
    get:
      summary: Lista as competências
      tags:
        - Content
      description: |-
        🔓 **Público** — não precisa de login.

        Retorna as habilidades exibidas no portfólio.
      responses:
        "200":
          description: Lista de competências.
          content:
            application/json:
              schema:
                description: Lista de competências.
                type: object
                required:
                  - revision
                  - updatedAt
                  - data
                properties:
                  revision:
                    type: integer
                    description: Versão atual dos dados. Use em `expectedRevision` ao salvar.
                    example: 4
                  updatedAt:
                    type: string
                    format: date-time
                    description: Data da última alteração.
                    example: 2025-01-15T12:00:00.000Z
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/def-4"
                    example:
                      - id: automacao-testes
                        title: Automação de testes
                        description: Criação de testes automatizados para web e API.
                        tags:
                          - Playwright
                          - TypeScript
    put:
      summary: Salva a lista de competências
      tags:
        - Content
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Envia a lista **inteira** de competências: itens que não estiverem na
        lista serão removidos. Dica: faça o GET, ajuste a lista e envie de
        volta. Antes de enviar, consulte o GET correspondente e use o valor
        `revision` retornado em `expectedRevision`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Versão atual + lista completa de itens.
              required:
                - expectedRevision
                - data
              additionalProperties: false
              properties:
                expectedRevision:
                  type: integer
                  minimum: 1
                  description: Versão atual dos dados. Use o valor `revision` retornado pelo
                    último GET.
                  example: 4
                data:
                  type: array
                  maxItems: 100
                  description: Lista completa. Ela **substitui** a lista atual.
                  items:
                    $ref: "#/components/schemas/def-4"
            example:
              expectedRevision: 4
              data:
                - id: automacao-testes
                  title: Automação de testes
                  description: Criação de testes automatizados para web e API.
                  tags:
                    - Playwright
                    - TypeScript
        description: Versão atual + lista completa de itens.
      security: *a1
      responses:
        "200":
          description: Salvo. Retorna o portfólio completo com a nova `revision`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-18"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "409":
          description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
            existe). Faça um novo GET e tente de novo."
          content:
            application/json:
              schema:
                description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
                  existe). Faça um novo GET e tente de novo."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: REVISION_CONFLICT
                  message:
                    type: string
                    description: Explicação do erro.
                    example: O portfólio foi alterado depois que esta edição começou. Recarregue a
                      revisão atual antes de salvar.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      expectedRevision: 3
                      currentRevision: 4
  /api/v1/education:
    get:
      summary: Lista formação e cursos
      tags:
        - Content
      description: |-
        🔓 **Público** — não precisa de login.

        Retorna a formação acadêmica e os cursos.
      responses:
        "200":
          description: Lista de formações e cursos.
          content:
            application/json:
              schema:
                description: Lista de formações e cursos.
                type: object
                required:
                  - revision
                  - updatedAt
                  - data
                properties:
                  revision:
                    type: integer
                    description: Versão atual dos dados. Use em `expectedRevision` ao salvar.
                    example: 4
                  updatedAt:
                    type: string
                    format: date-time
                    description: Data da última alteração.
                    example: 2025-01-15T12:00:00.000Z
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/def-7"
                    example:
                      - id: curso-exemplo
                        title: Curso de Teste de Software
                        institution: Instituição Exemplo
                        status: Concluído
                        period: "2023"
                        description: Fundamentos de testes, técnicas e boas práticas.
                        href: https://exemplo.com/curso
    put:
      summary: Salva formação e cursos
      tags:
        - Content
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Envia a lista **inteira** de formações e cursos: itens que não estiverem
        na lista serão removidos. Dica: faça o GET, ajuste a lista e envie de
        volta. Antes de enviar, consulte o GET correspondente e use o valor
        `revision` retornado em `expectedRevision`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Versão atual + lista completa de itens.
              required:
                - expectedRevision
                - data
              additionalProperties: false
              properties:
                expectedRevision:
                  type: integer
                  minimum: 1
                  description: Versão atual dos dados. Use o valor `revision` retornado pelo
                    último GET.
                  example: 4
                data:
                  type: array
                  maxItems: 100
                  description: Lista completa. Ela **substitui** a lista atual.
                  items:
                    $ref: "#/components/schemas/def-7"
            example:
              expectedRevision: 4
              data:
                - id: curso-exemplo
                  title: Curso de Teste de Software
                  institution: Instituição Exemplo
                  status: Concluído
                  period: "2023"
                  description: Fundamentos de testes, técnicas e boas práticas.
                  href: https://exemplo.com/curso
        description: Versão atual + lista completa de itens.
      security: *a1
      responses:
        "200":
          description: Salvo. Retorna o portfólio completo com a nova `revision`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-18"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "409":
          description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
            existe). Faça um novo GET e tente de novo."
          content:
            application/json:
              schema:
                description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
                  existe). Faça um novo GET e tente de novo."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: REVISION_CONFLICT
                  message:
                    type: string
                    description: Explicação do erro.
                    example: O portfólio foi alterado depois que esta edição começou. Recarregue a
                      revisão atual antes de salvar.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      expectedRevision: 3
                      currentRevision: 4
...
servers:
  - url: http://localhost:3333
    description: Desenvolvimento local
...